requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt==4.0.1
pytest
pytest-asyncio
httpx
pytest-mock
python-dotenv
aiofiles
